# swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0",
        "description": "A sample API"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "tags": [
        {
            "name": "test",
            "description": "test endpoint"
        },
        {
            "name": "auth",
            "description": "Authentication related endpoints"
        },
        {
            "name": "analyze",
            "description": "Analyze related endpoints"
        },
        {
            "name": "admin",
            "description": "Admin related endpoints"
        }
    ],
    "paths": {
        "/test": {
            "get": {
                "tags": ["test"],
                "summary": "Test endpoint",
                "description": "Test endpoint for testing",
                "responses": {
                    "200": {
                        "description": "Test response"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": ["auth"],
                "summary": "Google 로그인 및 회원가입",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "google_access_token": { "type": "string" },
                                    "phone_number": { "type": "string" },
                                    "source_path": { "type": "string" }
                                },
                                "required": ["google_access_token"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "access_token": { "type": "string" },
                                        "refresh_token": { "type": "string" },
                                        "user": {
                                            "type": "object",
                                            "properties": {
                                                "id": { "type": "string" },
                                                "email": { "type": "string" },
                                                "phone_number": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing or invalid input"
                    }
                }
            }
        },
        "/auth/token": {
            "post": {
                "tags": ["auth"],
                "summary": "Access token 재발급",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "refresh_token": { "type": "string" }
                                },
                                "required": ["refresh_token"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token refreshed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "access_token": { "type": "string" },
                                        "refresh_token": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid or expired refresh token"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": ["auth"],
                "summary": "로그아웃 (refresh token 삭제)",
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "204": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Invalid or missing access token"
                    }
                }
            }
        },
        "/analyze/upload": {
            "post": {
                "tags": ["analyze"],
                "summary": "Upload PDF and start analysis",
                "description": "Uploads a PDF file, stores it on S3, and triggers Gemini LLM to extract reagent info.",
                "security": [{ "bearerAuth": [] }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "PDF file to analyze"
                                    }
                                },
                                "required": ["file"]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "session_id": {
                                            "type": "string",
                                            "example": "a3f4d91c-7e9a-4db1-8f64-0012ef2d3c3d"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing file or invalid input"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/analyze/result/{session_id}": {
            "get": {
                "tags": ["analyze"],
                "summary": "Get analysis result by session ID",
                "description": "Returns the extracted reagent info for a given session. If not authenticated, catalog info is hidden.",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "Result found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "session_id": { "type": "string" },
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "processing",
                                                "done",
                                                "failed"
                                            ]
                                        },
                                        "results": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "company": {
                                                        "type": "string"
                                                    },
                                                    "reagent": {
                                                        "type": "string"
                                                    },
                                                    "catalog": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found"
                    }
                }
            }
        },
        "/analyze/result/{session_id}/csv": {
            "get": {
                "tags": ["analyze"],
                "summary": "Download CSV of extracted results",
                "description": "Returns extracted result as CSV file. Requires authentication for catalog field.",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "security": [{ "bearerAuth": [] }],
                "responses": {
                    "200": {
                        "description": "CSV file stream",
                        "content": {
                            "text/csv": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Session not found or no results"
                    },
                    "409": {
                        "description": "Result not ready yet"
                    }
                }
            }
        },
        "/admin/summary": {
            "get": {
                "tags": ["admin"],
                "summary": "관리자 요약 정보 조회",
                "description": "전체 사용자 목록, 논문 목록 및 추출된 시약 결과를 관리자 권한으로 조회합니다.",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "요약 정보 조회 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "email": {
                                                        "type": "string"
                                                    },
                                                    "phone_number": {
                                                        "type": "string"
                                                    },
                                                    "source_path": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        },
                                        "papers": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "paper_name": {
                                                        "type": "string"
                                                    },
                                                    "session_id": {
                                                        "type": "string"
                                                    },
                                                    "results": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "object",
                                                            "properties": {
                                                                "company": {
                                                                    "type": "string"
                                                                },
                                                                "reagent": {
                                                                    "type": "string"
                                                                },
                                                                "catalog": {
                                                                    "type": "string"
                                                                }
                                                            }
                                                        }
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "관리자 권한 없음",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": { "type": "integer" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}